coalesce hover svg logic
